+++ harness/wazero/binding.c
@@ -1,0 +1,5 @@
+#include <node_api.h>
+
+extern napi_value GoInit(napi_env env, napi_value exports);
+
+NAPI_MODULE(NODE_GYP_MODULE_NAME, GoInit)